requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
azure-storage-blob
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.3
multitasking==0.0.9
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
yfinance==0.1.59
cryptography==3.1.1